yuna 0621 09:35

+++ src/main/java/com/ball/service/GroupService.java
@@ -10,7 +10,7 @@ public interface GroupService {
     public GroupVO get(Long group_id);        //하나만 조회
     public List<GroupVO> allRead(Criteria cri);    //전체 가져오기
     public void modify (GroupVO group);     //수정
-    public Long remove(Long group_id);      //삭제
+    public int remove(Long group_id);      //삭제
 
     public int getTotal(Criteria cri);
